in product service, moved all price related classes to one package with a single publically accessible class
