+++ product-catalogue-service/src/main/java/com/globomart/productcatalogue/price/PriceCommandBuilder.java
@@ -1,9 +1,7 @@
-package com.globomart.productcatalogue.service.price;
+package com.globomart.productcatalogue.price;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
-
-import com.globomart.productcatalogue.client.PriceClient;
 
 /**
  * Builder class to create and initialize a hystrix command implementation instance to fetch price from price-service via Feign client.
